+++ relayer/workers/beefyrelayer/beefy-ethereum-writer.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/snowfork/polkadot-ethereum/relayer/chain/ethereum"
 	"github.com/snowfork/polkadot-ethereum/relayer/contracts/lightclientbridge"
-	"github.com/snowfork/polkadot-ethereum/relayer/store"
+	"github.com/snowfork/polkadot-ethereum/relayer/workers/beefyrelayer/store"
 )
 
 type BeefyEthereumWriter struct {
