+++ relayer/workers/beefyrelayer/store/store.go
@@ -153,6 +153,8 @@ func (d *Database) Stop() {
 func (d *Database) writeLoop(ctx context.Context) error {
 	var mutex = &sync.Mutex{}
 
+	fmt.Println("d.messages")
+	fmt.Println(d.messages)
 	for {
 		select {
 		case <-ctx.Done():
